backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiomysql
pydantic-settings
alembic
python-jose[cryptography]
passlib[bcrypt]
argon2-cffi
python-multipart
asyncpg
ujson
gunicorn
psycopg2-binary
aiofiles
python-dotenv
greenlet
scikit-learn
pandas
numpy
sentence-transformers
torch
